// src/services/types/alpaca.ts
export type AlpacaPosition = {
	asset_id: string;
	symbol: string;
	exchange: string;
	asset_class: string;
	avg_entry_price: string;
	qty: string;
	qty_available: string;
	side: string;
	market_value: string;
	cost_basis: string;
	unrealized_pl: string;
	unrealized_plpc: string;
	unrealized_intraday_pl: string;
	unrealized_intraday_plpc: string;
	current_price: string;
	lastday_price: string;
	change_today: string;
	asset_marginable: boolean;
};

export type AlpacaAccount = {
	id: string;
	admin_configurations: object;
	user_configurations: object;
	account_number: string;
	status: string;
	crypto_status: string;
	currency: string;
	buying_power: string;
	regt_buying_power: string;
	daytrading_buying_power: string;
	options_buying_power: string;
	effective_buying_power: string;
	non_marginable_buying_power: string;
	bod_dtbp: number;
	cash: string;
	accrued_fees: string;
	pending_transfer_in: string;
	portfolio_value: string;
	pattern_day_trader: boolean;
	trading_blocked: boolean;
	transfers_blocked: boolean;
	account_blocked: boolean;
	created_at: string;
	trade_suspended_by_user: boolean;
	multiplier: string;
	shorting_enabled: boolean;
	equity: string;
	last_equity: string;
	long_market_value: string;
	short_market_value: string;
	position_market_value: string;
	initial_margin: string;
	maintenance_margin: string;
	last_maintenance_margin: string;
	sma: string;
	daytrade_count: number;
	balance_asof: string;
	crypto_tier: number;
	options_trading_level: number;
	intraday_adjustments: string;
};